Fixing ImageListManager.Clear()
Originally image lists wouldn't be replaced when Clear() was called.
This was partly because the original implementation only worked for
ListViews, but was also partly because IconList.SmallIL_ID and
IconList.LargeIL_ID were both set to new image lists, causing the later
code in SetIcon() not to trigger. Instead we set these values to false.

Also some cleanup of comments in SetIcon()